prevent possible long/32bit problems

+++ include/tiles/fixed/fixed_geometry.h
@@ -16,7 +16,7 @@
 
 namespace tiles {
 
-using fixed_coord_t = signed long long int;
+using fixed_coord_t = int64_t;
 
 using fixed_xy = boost::geometry::model::d2::point_xy<fixed_coord_t>;
 
@@ -30,12 +30,12 @@ using fixed_ring = fixed_simple_polygon::ring_type;
 
 constexpr fixed_coord_t kFixedCoordMin = 0;
 constexpr fixed_coord_t kFixedCoordMax = proj::map_size(kMaxZoomLevel) - 1;
-constexpr fixed_coord_t kFixedCoordMagicOffset = kFixedCoordMax / 2ul;
+constexpr fixed_coord_t kFixedCoordMagicOffset = kFixedCoordMax / 2ULL;
 
-constexpr auto kFixedDefaultZoomLevel = 20ul;
+constexpr auto kFixedDefaultZoomLevel = 20ULL;
 static_assert(kFixedDefaultZoomLevel <= kMaxZoomLevel, "invalid default zoom");
 
-using fixed_delta_t = signed long long int;
+using fixed_delta_t = int64_t;
 
 using fixed_null = std::monostate;
 using fixed_point = boost::geometry::model::multi_point<fixed_xy>;
